requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
Pillow
pyinstaller
